change type of variable

+++ src/taskOne/Main.java
@@ -6,7 +6,7 @@ public class Main {
 
 	public static void main(String[] args) {
 		
-		int result = 1;
+		long result = 1;
 
 		Scanner sc = new Scanner(System.in);
